Add lifts.
Lifts act like passages, but have an extra effect in a gravity field. They allow players in or directly above them to fly without activating their jetpack. Flying in a lift without a jetpack is limited to sprinting speed (7m/s), and the player's feet remain pointed down relative to the gravity field. As soon as the player leaves the lift, they return to walking.

Lifts do not use energy. Instead, they only work if they are in a net gravity field of at least 0.5G; they are assumed to be powered by the gravity generator.

No actual lift block is included.